Refactor: Remove unused components and contexts; reintroduce ToggleSwitch and SoundButton with updated implementations
- Deleted ToggleSwitch, HotkeyModal, SoundButton, SoundGrid, AudioContext, and SettingsContext components.
- Added new implementations for ToggleSwitch, SoundButton, SoundGrid, AudioContext, and SettingsContext.
- Introduced new icons: StopIcon, Windows, Multi, Repeat, Mute, and Volume.
- Updated imports and component structure for better organization and maintainability.
