+++ src/components/icons/index.ts
@@ -1,0 +1,6 @@
+export { Volume } from '@/components/icons/volume/volume';
+export { Mute } from '@/components/icons/volume/mute';
+export { StopIcon } from '@/components/icons/control/stop';
+export { Windows } from '@/components/icons/control/windows';
+export { Multi } from '@/components/icons/toggle/multi';
+export { Repeat } from '@/components/icons/toggle/repeat';
