+++ src/components/modals/hotkey.tsx
@@ -1,13 +1,11 @@
 import React, { useCallback, useEffect, useState } from 'react';
 import {
   Dialog,
-  DialogClose,
   DialogContent,
   DialogDescription,
   DialogFooter,
   DialogHeader,
   DialogTitle,
-  DialogTrigger,
 } from "@/components/ui/dialog"
 import { Button } from '@/components/ui/button';
